+++ src/git.rs
@@ -6,20 +6,42 @@
 
 use std::path::Path;
 use std::process::Command;
-use std::result;
+use thiserror::Error;
 
 use log::debug;
 
-type Result = result::Result<(), String>;
-
-pub trait GitWrapper {
-    fn git_version(&self) -> Result;
-    fn git_clone_mirror(&self, origin: &str, repo_dir: &Path) -> Result;
-    fn git_update_mirror(&self, origin: &str, repo_dir: &Path) -> Result;
-    fn git_push_mirror(&self, dest: &str, repo_dir: &Path, refspec: &Option<Vec<String>>)
-        -> Result;
+/// An error occuring during git command execution
+#[derive(Debug, Error)]
+pub enum GitError {
+    #[error("Command {cmd:?} failed with error: {err}")]
+    CommandError { cmd: Command, err: std::io::Error },
+    #[error("Command {cmd:?} failed with exit code: {code}, Stderr: {stderr}")]
+    GitCommandError {
+        code: i32,
+        stderr: String,
+        cmd: Command,
+    },
 }
 
+/// Common interface to different git backends
+/// - [x] git command line
+/// - [ ] libgit2
+/// - [ ] gitoxide
+///
+pub trait GitWrapper {
+    /// Get the git version
+    fn git_version(&self) -> Result<(), GitError>;
+    fn git_clone_mirror(&self, origin: &str, repo_dir: &Path) -> Result<(), GitError>;
+    fn git_update_mirror(&self, origin: &str, repo_dir: &Path) -> Result<(), GitError>;
+    fn git_push_mirror(
+        &self,
+        dest: &str,
+        repo_dir: &Path,
+        refspec: &Option<Vec<String>>,
+    ) -> Result<(), GitError>;
+}
+
+/// Git command line wrapper
 pub struct Git {
     executable: String,
 }
@@ -35,7 +57,7 @@ impl Git {
         git
     }
 
-    fn run_cmd(&self, mut cmd: Command) -> Result {
+    fn run_cmd(&self, mut cmd: Command) -> Result<(), GitError> {
         debug!("Run command: {:?}", cmd);
         match cmd.output() {
             Ok(o) => {
@@ -50,28 +72,27 @@ impl Git {
                 if o.status.success() {
                     Ok(())
                 } else {
-                    Err(format!(
-                        "Command {:?} failed with exit code: {}, Stderr: {}",
+                    Err(GitError::GitCommandError {
                         cmd,
-                        o.status.code().unwrap_or_default(),
-                        stderr
-                    ))
+                        code: o.status.code().unwrap_or_default(),
+                        stderr,
+                    })
                 }
             }
-            Err(e) => Err(format!("Command {:?} failed with error: {}", cmd, e)),
+            Err(e) => Err(GitError::CommandError { cmd, err: e }),
         }
     }
 }
 
 impl GitWrapper for Git {
-    fn git_version(&self) -> Result {
+    fn git_version(&self) -> Result<(), GitError> {
         let mut cmd = self.git_base_cmd();
         cmd.arg("--version");
 
         self.run_cmd(cmd)
     }
 
-    fn git_clone_mirror(&self, origin: &str, repo_dir: &Path) -> Result {
+    fn git_clone_mirror(&self, origin: &str, repo_dir: &Path) -> Result<(), GitError> {
         let mut clone_cmd = self.git_base_cmd();
         clone_cmd
             .args(&["clone", "--mirror"])
@@ -81,7 +102,7 @@ impl GitWrapper for Git {
         self.run_cmd(clone_cmd)
     }
 
-    fn git_update_mirror(&self, origin: &str, repo_dir: &Path) -> Result {
+    fn git_update_mirror(&self, origin: &str, repo_dir: &Path) -> Result<(), GitError> {
         let mut set_url_cmd = self.git_base_cmd();
         set_url_cmd
             .current_dir(repo_dir)
@@ -103,7 +124,7 @@ impl GitWrapper for Git {
         dest: &str,
         repo_dir: &Path,
         refspec: &Option<Vec<String>>,
-    ) -> Result {
+    ) -> Result<(), GitError> {
         let mut push_cmd = self.git_base_cmd();
         push_cmd.current_dir(repo_dir);
         push_cmd.arg("push");